requirements.txt
python-dotenv==1.0.1
confluent-kafka==2.4.0
pyspark==3.5.1
delta-spark==3.2.0